refactor(render_window_serializer): rename to serializers
This name fits better since it contains all kinds of serializers.
